controller_interface: Add MultiInterfaceController
MultiInterfaceController: Subclass of ControllerBase which allows to claim
resources from up to four different hardware interfaces.

It's similar in spirit to the single-interface Controller class, in that
hardware interfaces are specified as template parameters. One difference is that
the custom controller initialization init() methods are passed a RobotHW*
parameter containing only the hardware interfaces requested by the controller
(as opposed to the actual hardware interface for Controller, or a RobotHW*
representing the _entire_ robot for ControllerBase).

Requested hardware interfaces are required by default, but can be made optional
through a constructor flag.