Handle building without GMock
These macros will not fail if GMock is not available at build time, and
will not produce a target under these circumstances. However, the
subsequent CMake lines are assuming that the target will exist, and
cause CMake to fail if not.

This change makes the packages tolerant to the absence of GMock at build
time.