require REQUIRED build deps as build_depend
This package is not very critical and it is unclear
what it should do at all when testing is disabled.
But as the CMakeLists.txt requires these packages during either build,
they are NOT just test dependencies.

Required to make use of added functions to exclude test depends in builds
https://github.com/ros-infrastructure/bloom/pull/649